+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## 1.21.2 (PENDING)
 
+### Changes
+* Deprecated internal method `Normalizer#normalize(String, bool)`.
+
 ### Improvements
 * When pretty-printing, if there are consecutive text nodes (via DOM manipulation), the non-significant whitespace between them will be collapsed. [#2349](https://github.com/jhy/jsoup/pull/2349).
 * Updated `Connection.Response#statusMessage()` to return a simple loggable string message (e.g. "OK") when using the `HttpClient` implementation, which doesn't otherwise return any server-set status message. [#2356](https://github.com/jhy/jsoup/issues/2346)